+++ cmp/zoneGrid/ZoneGridModel.ts
@@ -583,7 +583,7 @@ export class ZoneGridModel extends HoistModel {
             const colId = it.field;
             subFields.push({
                 colId,
-                label: it.showLabel ? labelRenderers[colId] ?? true : false,
+                label: it.showLabel ? (labelRenderers[colId] ?? true) : false,
                 position: 'top'
             });
         });
@@ -591,7 +591,7 @@ export class ZoneGridModel extends HoistModel {
             const colId = it.field;
             subFields.push({
                 colId,
-                label: it.showLabel ? labelRenderers[colId] ?? true : false,
+                label: it.showLabel ? (labelRenderers[colId] ?? true) : false,
                 position: 'bottom'
             });
         });
@@ -608,7 +608,7 @@ export class ZoneGridModel extends HoistModel {
             // Controlled properties
             field: isLeft ? 'left_column' : 'right_column',
             align: isLeft ? 'left' : 'right',
-            isLeftZoneColumn: isLeft,
+            isTreeColumn: gridModel?.treeMode && isLeft,
             flex: overrideSpec.width ? null : isLeft ? 2 : 1,
             renderer: (value, context) => zoneGridRenderer(value, context, isLeft),
             rendererIsComplex: true,